+++ tests/testthat/test-auth.R
@@ -1,9 +1,5 @@
 context("Authentication")
 
-test_that("On package load, the session_store exists", {
-    expect_true(is.environment(session_store))
-})
-
 test_that("login checks for email and password before POSTing", {
     expect_error(crunchAuth(email=NULL),
         "Must supply the email address associated with your crunch.io account")
@@ -31,15 +27,6 @@ with_mock_crunch({
 
 if (run.integration.tests) {
     with(test_options, {
-        test_that("login works if crunch is running", {
-            deleteSessionInfo()
-            suppressMessages(login())
-                expect_true("root" %in% ls(envir=session_store))
-                expect_true(is.authenticated())
-            logout()
-            expect_false(is.authenticated())
-        })
-
         test_that("crunchAuth succeeds when it should and not when it shouldn't", {
             logout()
             expect_error(crunchAuth("lkjasdfksdfkjhl", password="w23nrnsod"),