+++ vignettes/crunch/0/api/datasets/d2d639/batches-c322df-POST.R
@@ -1,0 +1,29 @@
+structure(list(url = "/api/datasets/d2d639/batches/", status_code = 202L, 
+    headers = structure(list(allow = "GET, HEAD, OPTIONS, POST", 
+        `content-encoding` = "gzip", `content-type` = "application/json;charset=utf-8", 
+        date = "Thu, 10 May 2018 21:54:31 GMT", location = "/api/datasets/d2d639/batches/1/", 
+        server = "nginx", `set-cookie` = "REDACTED", vary = "Cookie, Accept-Encoding", 
+        `x-timing` = "", `content-length` = "180", connection = "keep-alive"), .Names = c("allow", 
+    "content-encoding", "content-type", "date", "location", "server", 
+    "set-cookie", "vary", "x-timing", "content-length", "connection"
+    ), class = c("insensitive", "list")), all_headers = list(
+        structure(list(status = 202L, version = "HTTP/1.1", headers = structure(list(
+            allow = "GET, HEAD, OPTIONS, POST", `content-encoding` = "gzip", 
+            `content-type` = "application/json;charset=utf-8", 
+            date = "Thu, 10 May 2018 21:54:31 GMT", location = "/api/datasets/d2d639/batches/1/", 
+            server = "nginx", `set-cookie` = "REDACTED", vary = "Cookie, Accept-Encoding", 
+            `x-timing` = "", `content-length` = "180", connection = "keep-alive"), .Names = c("allow", 
+        "content-encoding", "content-type", "date", "location", 
+        "server", "set-cookie", "vary", "x-timing", "content-length", 
+        "connection"), class = c("insensitive", "list"))), .Names = c("status", 
+        "version", "headers"))), cookies = structure(list(domain = ".crunch.io", 
+        flag = TRUE, path = "/", secure = FALSE, expiration = structure(1557525271, class = c("POSIXct", 
+        "POSIXt")), name = "token", value = "REDACTED"), .Names = c("domain", 
+    "flag", "path", "secure", "expiration", "name", "value"), row.names = c(NA, 
+    -1L), class = "data.frame"), content = charToRaw("{\"element\": \"shoji:view\", \"self\": \"/api/datasets/d2d639/batches/\", \"value\": \"/api/progress/\"}"), 
+    date = structure(1525989271, class = c("POSIXct", "POSIXt"
+    ), tzone = "GMT"), times = structure(c(0, 4.3e-05, 4.5e-05, 
+    0.000126, 0.204895, 0.204973), .Names = c("redirect", "namelookup", 
+    "connect", "pretransfer", "starttransfer", "total"))), .Names = c("url", 
+"status_code", "headers", "all_headers", "cookies", "content", 
+"date", "times"), class = "response")